chore: check on installation type now as single module
