+++ extensions/interfaces/iinstallation-type.ts
@@ -1,0 +1,4 @@
+export interface IInstallationType {
+ isUpdate: boolean;
+ isFirstInstall: boolean;
+}
